Focusing Hard
Could definitely be improved by some logic to wait for the prompt that draws focus to appear before refocusing on the 2FA box, but that seemed excessive (and would probably break if USCIS updated their website).

Could do "wait for it, OR after 500ms" or something, but that again seems a bit excessive and I've already dedicated 5 whole minutes to setting this up to solve a stupid slight inconvenience, which frankly is 5 minutes more than I should have.